python/requirements.txt
# Excel Automation Framework Requirements

# Configuration
pyyaml>=6.0.1

# CLI
click>=8.1.7
rich>=13.7.0

# Scheduling
schedule>=1.2.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Type checking
mypy>=1.7.1
types-PyYAML>=6.0.12

# Code quality
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

# Development
ipython>=8.18.1
